+++ CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- add support for computed fields in graph queries
+
 ### Changes
+
+- BREAKING: make `MEX_EXTRACTED_PRIMARY_SOURCE` an instance of its own class
+  instead of ExtractedPrimarySource in order to set static provenance identifiers
 
 ### Deprecated